+++ console/packages/api-client/src/api.ts
@@ -18,9 +18,6 @@ export * from './api/v1alpha1-attachment-api';
 export * from './api/v1alpha1-attachment-reference-api';
 export * from './api/v1alpha1-attachment-relation-api';
 export * from './api/v1alpha1-episode-collection-api';
-export * from './api/v1alpha1-file-api';
-export * from './api/v1alpha1-file-relation-api';
-export * from './api/v1alpha1-folder-api';
 export * from './api/v1alpha1-indices-api';
 export * from './api/v1alpha1-plugin-api';
 export * from './api/v1alpha1-subject-api';
