+++ console/packages/api-client/src/models/index.ts
@@ -3,17 +3,11 @@ export * from './attachment-reference';
 export * from './attachment-relation';
 export * from './author';
 export * from './batch-matching-episode-attachment';
+export * from './batch-matching-subject-episodes-attachment';
 export * from './config-map';
 export * from './episode';
 export * from './episode-collection';
 export * from './episode-resource';
-export * from './file-batch-request';
-export * from './file-entity';
-export * from './file-hint';
-export * from './file-hints';
-export * from './file-relations';
-export * from './folder';
-export * from './model-file';
 export * from './paging-wrap';
 export * from './plugin';
 export * from './plugin-load-location';
@@ -24,7 +18,6 @@ export * from './subject-hint';
 export * from './subject-hints';
 export * from './subject-relation';
 export * from './subject-sync';
-export * from './subtitle';
 export * from './task-entity';
 export * from './update-user-request';
 export * from './user';
