+++ console/packages/api-client/src/models/paging-wrap.ts
@@ -43,17 +43,17 @@ export interface PagingWrap {
 	 */
 	items: Array<object>;
 	/**
-	 * Indicates whether current page is the first page.
-	 * @type {boolean}
-	 * @memberof PagingWrap
-	 */
-	firstPage: boolean;
-	/**
 	 * Indicates whether current page is the last page.
 	 * @type {boolean}
 	 * @memberof PagingWrap
 	 */
 	lastPage: boolean;
+	/**
+	 * Indicates whether current page is the first page.
+	 * @type {boolean}
+	 * @memberof PagingWrap
+	 */
+	firstPage: boolean;
 	/**
 	 * Indicates whether current page has previous page.
 	 * @type {boolean}
